photography-master-app_3/api/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.4
python-dotenv==1.0.0
openai==1.12.0
SpeechRecognition==3.14.3
gTTS==2.5.4
pydub==0.25.1
pygame==2.6.1
